Adapt to removal of URLs list from StartupBrowserCreatorImpl::Launch()
We no longer have the |urls_to_open| parameter there so we need to
split the functionality of our override for the Launch() method in
two parts:

  A. Switching to the Tor profile when specificied via cmdline args.
  B. Fetch URL from cmdline args if MS-Edge protocol is used.

For (A) we can keep using the override for the Launch() method and
simply make sure we use the Tor profile.

For (B) we override StartupTabProviderImpl::GetCommandLineTabs() so
that we include the URL using MS-Edge protocol when specified via
cmdline args, which is not something we can continue doing in the
Launch() method as no URLs list will be available there anymore.

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/33c1710d146c19f64884bc26e43312e07397ce01

commit 33c1710d146c19f64884bc26e43312e07397ce01
Author: Hidehiko Abe <hidehiko@chromium.org>
Date:   Tue Aug 24 01:11:56 2021 +0000

    Move GetURLsFromCommandLine to StartupTabProvider.

    This is refactoring to prepare Lacros launching URL support.

    Bug: 1236859